pyproject.toml
[project]
name = "voice-assistant"
version = "0.1.0"
description = "Local voice assistant — Ubuntu 24.04, RTX 3090"
requires-python = ">=3.12"
dependencies = [
    "numpy==2.4.4",
    "sounddevice==0.4.7",
    "pynvml==13.0.1",
    "aiohttp==3.13.5",
    "pyyaml==6.0.3",
    "jsonschema==4.26.0",
    "langdetect==1.0.9",
    "torchaudio==2.11.0",
    "omnivoice==0.1.3",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "ruff"]

[build-system]
requires = ["setuptools>=70"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"
